requirements.txt
python-dotenv
openai
pydantic
streamlit
